requirements.txt
numpy==1.23.2
Pillow==9.2.0
tqdm==4.64.0